file open and stat impl.

+++ src/java/src/main/java/com/fastken/fcfs/FCFSFile.java
@@ -42,7 +42,7 @@ public class FCFSFile {
         this.fd = fd;
     }
 
-    public long getFD() {
+    public int getFD() {
         return this.fd;
     }
